+++ src/schematics/storybook/storybook.ts
@@ -23,7 +23,14 @@ export default function(options: StorybookSchema): Rule {
             '@storybook/addon-options',
             '@storybook/angular',
             '@storybook/components',
+            '@types/storybook__addon-actions',
+            '@types/storybook__addon-links',
+            '@types/storybook__addon-notes',
+            '@types/storybook__addon-options',
+            '@types/webpack',
             'babel-loader',
+            'css-loader',
+            'style-loader',
         ]),
         renderTemplates('./files', './', {
             params: {
@@ -35,7 +42,7 @@ export default function(options: StorybookSchema): Rule {
         conditional(!options.skipScript, [
             addScript(
                 'storybook',
-                `start-storybook --port ${options.port} --config-dir ${options.config}`,
+                `start-storybook --config-dir ${options.config} --port ${options.port}`,
             ),
         ]),
         conditional(!options.skipInstall, [installDependencies()]),